extracted test_helpers.rs
